% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qrp_qr.R
\name{B_orth}
\alias{B_orth}
\title{B-spline orthogonalization}
\usage{
B_orth(LLI, degree, bkn, crep_input = NULL, is_approx = FALSE)
}
\arguments{
\item{is_approx}{}
}
\value{
This function returns a \code{list} including: 
\itemize{
 \item b_function
 \item z
}
}
\description{
B-spline orthogonalization
}
