% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantRegGroups.R
\name{bSpline}
\alias{bSpline}
\title{B-spline orthogonalization}
\usage{
bSpline(LLI, degree, bkn, crep_input = NULL, is_approx = FALSE)
}
\arguments{
\item{is_approx}{}
}
\value{
This function returns a \code{list} including:
\itemize{
\item b_function
\item z
}
}
\description{
B-spline orthogonalization
}
