% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{orth_bspline}
\alias{orth_bspline}
\title{Orthogonalized B-splines}
\usage{
orth_bspline(
  knots,
  boundary_knots,
  degree,
  predictors = NULL,
  is_approx = FALSE
)
}
\arguments{
\item{knots}{Array. The knots that define the spline.}

\item{boundary_knots}{Array. The breakpoints that define the spline.}

\item{degree}{Integer. The degree of the piecewise polynomial.}

\item{predictors}{Array. The predictor variables with size p.}

\item{is_approx}{Boolean. The default is \code{FALSE}.}
}
\value{
Matrix Orthogonalized B-splines with column size p.
}
\description{
Orthogonalized B-splines
}
\examples{
p <- 30
total_knots <- 10
degree <- 3
boundaries <- c(0, 1)
x <- seq(from = 0, to = 1,length.out = total_knots)
knots <- x[2:(total_knots - 1)]
predictors <- runif(p, min = 0, max = 1)
bsplines <- orth_bspline(knots, boundaries, degree, predictors)
}
