% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qrglasso.R
\name{plot.qrglasso.predict}
\alias{plot.qrglasso.predict}
\title{Display the estimated coefficient functions}
\usage{
\method{plot}{qrglasso.predict}(x, ...)
}
\arguments{
\item{x}{An object of class \code{qrglasso.predict} for the \code{plot} method}

\item{...}{Not used directly}
}
\value{
\code{NULL}
}
\description{
Display the estimated coefficient functions by BIC
}
\examples{
set.seed(123)
n <- 100
p <- 5
L <- 5
Y <- matrix(rnorm(n), n, 1)
W <- matrix(rnorm(n * p * (L - 1)), n, p * (L - 1))

result <- qrglasso(Y = Y, W = W, L = 5)
estimate <- predict(result, top_k = 2)
plot(estimate)

}
\seealso{
\code{\link{qrglasso}}
}
