% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{plot_bic_result}
\alias{plot_bic_result}
\title{Internal Function: Plot BIC Results w.r.t. lambda}
\usage{
plot_bic_result(data, variate)
}
\arguments{
\item{data}{Dataframe. A dataframe containing columns \code{lambda}, \code{bic}.}

\item{variate}{Character. A character representing the title.}
}
\value{
A ggplot object.
}
\description{
Internal function to plot BIC results with respect to lambda using ggplot2.
}
\keyword{internal}
