% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{plot_sequentially}
\alias{plot_sequentially}
\title{Internal Function: Plot Sequentially}
\usage{
plot_sequentially(objs)
}
\arguments{
\item{objs}{List. Valid ggplot2 objects to be plotted sequentially.}
}
\value{
\code{NULL}.
}
\description{
Internal function to plot ggplot2 objects sequentially.
}
\keyword{internal}
