% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qrglasso.R
\name{qrglasso}
\alias{qrglasso}
\title{Adaptively weighted group Lasso}
\usage{
qrglasso(
  Y,
  W,
  L,
  omega = NULL,
  tau = 0.5,
  qn = 1,
  lambda = NULL,
  maxit = 1000,
  thr = 1e-04
)
}
\arguments{
\item{Y}{A  \eqn{n \times 1} data matrix}

\item{W}{A  \eqn{n \times pL}B-spline matrix.}

\item{L}{The number of groups}

\item{omega}{A $p x 1$ weight matrix. Default value is NULL.}

\item{tau}{A numeric quantile of interest. Default value is 0.5.}

\item{qn}{A nuneric bound parameter for HDIC.  Default value is 1.}

\item{lambda}{A sequence of tuning parameters. Default value is NULL.}

\item{maxit}{The maximum number of iterations. Default value is 1000.}

\item{thr}{Threshold for convergence. Default value is \eqn{10^{-4}}.}
}
\value{
This function returns a \code{list} including:
\itemize{
\item{gamma}{A target estimate}
\item{xi}{An auxiliary estimate in the ADMM algorithm}
\item{phi}{An auxiliary estimate in the ADMM algorithm}
\item{BIC}{A sequence of BIC values w.r.t. different lambdas}
\item{lambda}{A sequence of tuning parameters used in the algorithm.}
\item{omega}{A $p x 1$ weight matrix used in the algorithm.}
}
}
\description{
Adaptively weighted group Lasso
}
