% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantRegGroups.R
\name{quantRegGroups}
\alias{quantRegGroups}
\title{Adaptively weighted group Lasso}
\usage{
quantRegGroups(
  Y,
  W,
  L,
  omega = NULL,
  tau,
  qn = 1,
  lambda = NULL,
  maxit = 1000,
  thr = 1e-04
)
}
\arguments{
\item{Y}{}
}
\value{
This function returns a \code{list} including:
\itemize{
\item gamma
\item xi
\item phi
\item BIC
\item lambda
\item omega
}
}
\description{
Adaptively weighted group Lasso
}
