% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{check_predict_parameters}
\alias{check_predict_parameters}
\title{Internal Function: Validate Parameters for Prediction with a \code{qrglasso} Object}
\usage{
check_predict_parameters(
  qrglasso_object,
  metric_type,
  top_k,
  degree,
  boundaries
)
}
\arguments{
\item{qrglasso_object}{A \code{qrglasso} class object.}

\item{metric_type}{Character. Metric type for gamma selection, e.g., \code{BIC}, \code{BIC-log}. Default is \code{BIC}.}

\item{top_k}{Integer. Top K estimated functions.}

\item{degree}{Integer. Degree of the piecewise polynomial.}

\item{boundaries}{Array. Two boundary points.}
}
\value{
\code{NULL}.
}
\description{
Internal function to validate parameters for predicting with a \code{qrglasso} class object.
}
\keyword{internal}
