% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{plot_bic_result}
\alias{plot_bic_result}
\title{Internal function: Plot BIC Results w.r.t. lambda}
\usage{
plot_bic_result(data, variate)
}
\arguments{
\item{data}{A dataframe contains columns \code{lambda}, \code{bic}}

\item{variate}{A character represent the title}
}
\value{
A ggplot object
}
\description{
Internal function: Plot BIC Results w.r.t. lambda
}
\keyword{internal}
