% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{plot_coefficient_function}
\alias{plot_coefficient_function}
\title{Internal function: Plot 2D fields for cross validation results}
\usage{
plot_coefficient_function(data, variate)
}
\arguments{
\item{data}{A dataframe contains columns \code{z}, \code{coef}}

\item{variate}{A character represent the title}
}
\value{
A ggplot object
}
\description{
Internal function: Plot 2D fields for cross validation results
}
\keyword{internal}
