% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{plot_coefficient_function}
\alias{plot_coefficient_function}
\title{Internal Function: Plot Coefficient Function}
\usage{
plot_coefficient_function(data, variate)
}
\arguments{
\item{data}{Dataframe. A dataframe containing columns \code{z}, \code{coefficient}.}

\item{variate}{Character. A character representing the title.}
}
\value{
A ggplot object.
}
\description{
Internal function to plot coefficient functions using ggplot2.
}
\keyword{internal}
