% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qrglasso.R
\name{predict}
\alias{predict}
\title{Predict the coefficient functions}
\usage{
predict(
  qrglasso_object,
  top_k = 5,
  degree = 2,
  boundaries = c(0, 1),
  is_approx = FALSE
)
}
\arguments{
\item{qrglasso_object}{An \code{qrglasso} class object.}

\item{top_k}{Integer. A matrix of the top K estimated functions. Default is 5.}

\item{degree}{Integer. Degree of the piecewise polynomial. Default is 2.}

\item{boundaries}{Array. Two boundary points. Default is c(0, 1).}

\item{is_approx}{Logical. If TRUE, the size of covariate indexes will be 1e6; otherwise, 1e4. Default is FALSE.}
}
\value{
A prediction matrix of Y at the new locations, x_new.
}
\description{
Predict the top-k coefficient functions
}
\examples{
set.seed(123)
n <- 100
p <- 5
L <- 5
Y <- matrix(rnorm(n), n, 1)
W <- matrix(rnorm(n * p * (L - 1)), n, p * (L - 1))

# Call qrglasso with default parameters
result <- qrglasso(Y = Y, W = W, L = 5)
estimate <- predict(result) 
print(dim(estimate))

}
\seealso{
\link{qrglasso}
}
